NAV2018 W1 CU07 package of the TransactionCurrency module
This is the unified diff of the TrancactionCurrency module, directly applicable to the NAV2018 W1 CU07 Rev. 23019 source tree.

The command to apply the diff is "git apply < nav-transactioncurrency.diff" in the root of the repository tree.
For splitting of the object files, use NAVObjectSplitter https://forum.mibuso.com/discussion/17715/nav-object-splitter-v3-0-0-0